refactor(engine-client): extract topics; add WS setClient for client swaps
EngineWebSocket previously froze its HoustonClient at construction, so a
later config swap (different baseUrl/token) would let REST calls hit the
new endpoint while the WS kept reconnecting against the old one. The new
setClient method swaps the client and forces a reconnect; existing topic
subscriptions and event handlers are preserved by the reconnect path.

Topic constants moved to their own file to keep ws.ts under the 200-line
limit after adding the new method.
